Fix: calculation of perspective multiple.
This incurs a change in the concept of perspective multiple, which is based on the vertical
field of view, and now can (/ has to) be parameterized on projection height, so as to remove
the Gfx::GetLogicalHeight() dependency from xr::Transforms.

In order to facilitate this, the perspective multiple is now cached separately (as calculating it
from tanHaflVerticalFov requires a division).

- xr::Camera was updated with the same changes;
- minor cleanup / removal of unused and / or not meaningful data members;