Changed the MacOS build to Intel-only.
The ARM64 dependencies are not trivial to build (brew has dropped support for the --universal switch), and since we no longer support Mac, we just want to be able to build and run what we have, without having to care about linker errors. Rosetta should help those that want to do this on M1.